Add validation

+++ shopping-list-backend/src/main/java/pl/edu/pwr/pastuszek/shoppinglistbackend/logic/service/MappedCrudService.java
@@ -5,10 +5,12 @@ import org.slf4j.Logger;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import pl.edu.pwr.pastuszek.shoppinglistbackend.logic.repositorie.BaseRepository;
 import pl.edu.pwr.pastuszek.shoppinglistbackend.logic.service.filter.CreatSpecifications;
 import pl.edu.pwr.pastuszek.shoppinglistbackend.model.dto.mapper.DTOMapper;
 import pl.edu.pwr.pastuszek.shoppinglistbackend.model.entity.Entity;
+import pl.edu.pwr.pastuszek.shoppinglistbackend.security.UserAuthentication;
 
 import java.util.List;
 import java.util.Map;
@@ -28,13 +30,19 @@ public abstract class MappedCrudService<T extends Entity, U, D>
 
     protected final DTOMapper<T, U, D> mapper;
 
-    protected MappedCrudService(BaseRepository<T> repository, Logger logger, DTOMapper<T, U, D> mapper, CreatSpecifications<T> creatSpecifications) {
-        super(repository, logger, creatSpecifications);
+    protected MappedCrudService(BaseRepository<T> repository,
+                                Logger logger,
+                                DTOMapper<T, U, D> mapper,
+                                CreatSpecifications<T> creatSpecifications,
+                                UserAuthentication userAuthentication
+    ) {
+        super(repository, logger, creatSpecifications, userAuthentication);
         this.mapper = mapper;
     }
 
     @Override
-    public Page<D> list(Map<String, String> params, Pageable pageable) {
+    public Page<D> list(Map<String, String> params, Pageable pageable) throws RuntimeException {
+        if(!isValidToList(params, pageable)) throw new AccessDeniedException("Access denied!");
         Page<T> entityPage = repository.findAll(creatSpecifications.creat(params), pageable);
 
         List<D> dtoList = entityPage.getContent().stream()
@@ -46,6 +54,7 @@ public abstract class MappedCrudService<T extends Entity, U, D>
 
     @Override
     public D getOne(UUID id) throws RuntimeException {
+        if(!isValidToGetOne(id)) throw new AccessDeniedException("Access denied!");
         var entity = repository
                 .findById(id)
                 .orElseThrow(() -> {
@@ -59,6 +68,7 @@ public abstract class MappedCrudService<T extends Entity, U, D>
 
     @Override
     public D add(U u) throws RuntimeException {
+        if(!isValidToAdd(u)) throw new AccessDeniedException("Access denied!");
         var entity = mapper.convertDtoToFullEntity(u);
         entity.setId(null);
         return mapper.convertEntityToDTO(repository.save(entity));
@@ -66,6 +76,7 @@ public abstract class MappedCrudService<T extends Entity, U, D>
 
     @Override
     public D update(UUID id, U u) throws RuntimeException {
+        if(!isValidToUpdate(id, u)) throw new AccessDeniedException("Access denied!");
         if (!repository.existsById(id)) {
             var e = new EntityNotFoundException("Entity of id " + id + " doesn't exist!");
             logger.error(e.getMessage(), e);
@@ -78,7 +89,14 @@ public abstract class MappedCrudService<T extends Entity, U, D>
     }
 
     @Override
-    public void delete(UUID id) {
+    public void delete(UUID id) throws RuntimeException{
+        if(!isValidToDelete(id)) throw new AccessDeniedException("Access denied!");
         repository.deleteById(id);
     }
+
+    protected abstract boolean isValidToList(Map<String, String> params, Pageable pageable);
+    protected abstract boolean isValidToGetOne(UUID id);
+    protected abstract boolean isValidToAdd(U u);
+    protected abstract boolean isValidToUpdate(UUID id, U u);
+    protected abstract boolean isValidToDelete(UUID id);
 }
